pyproject.toml
[project]
name = "european-strategy-consortium"
version = "0.1.0"
description = "Multi-agent system for strategic decision-making following European values"
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-mistralai>=0.1.0",
    "langchain-google-genai>=0.1.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
    "mistralai>=0.1.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "chromadb>=0.4.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.1",
    "black>=23.0",
    "mypy>=1.0",
    "flake8>=6.0",
    "isort>=5.12",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/consortium --cov-report=term-missing"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
